Serious bug in MQTTKit.m
Serious bug here, which tied up my server for a week.  initWithClientId takes BOOL cleanSession as parameter, which is almost always set to YES.  However, when it sends it to mosquito_new(), it sends the class property, self.cleanSession, which is not initialized, thus is always NO.  Using this code sets durable clients in the file, and causes the transmit queue to clog up waiting for those clients to sign in again and get their messages.